TroubleShooting 404 Error
After following above instructions, I ran into a 404 error which went as :

"If this is your site, make sure that the filename case matches the URL.
For root URLs (like http://example.com/) you must provide an index.html file."

This was because my app was using routing and used browserRouter to handle it. Legacy browsers don't support browser router. If a hashrouter is used itself, the problem resolves itself.